+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# v0.1.4
+* Fixed error when requirement file starts with wrong header level, like `##` instead of `#`.
+* Fixed `chk` command output for non-unique identifiers.
+
 # v0.1.3
 * Fixed errors output during repository reading process.